+++ src/main/java/com/quasiris/qsf/pipeline/filter/elastic/Elastic2SearchResultMappingTransformer.java
@@ -218,6 +218,9 @@ public class Elastic2SearchResultMappingTransformer implements SearchResultTrans
             }
 
             FacetValue facetValue = new FacetValue(key, count);
+            if(bucket.getCustomData() != null) {
+                facetValue.setCustomData(bucket.getCustomData());
+            }
             facetReseultCount = facetReseultCount + facetValue.getCount();
 
             facetFilterMapper.map(facetValue);